fix: extract backfillSteps into RunnerPoller+Backfill.swift, restore setDisplayState in-file (SW-R1002)
